Fully continuous curve + typing uniformity

+++ contracts/ReliquaryEvents.sol
@@ -2,24 +2,24 @@
 pragma solidity ^0.8.17;
 
 library ReliquaryEvents {
-    event CreateRelic(uint indexed pid, address indexed to, uint indexed relicId);
-    event Deposit(uint indexed pid, uint amount, address indexed to, uint indexed relicId);
-    event Withdraw(uint indexed pid, uint amount, address indexed to, uint indexed relicId);
-    event EmergencyWithdraw(uint indexed pid, uint amount, address indexed to, uint indexed relicId);
-    event Harvest(uint indexed pid, uint amount, address indexed to, uint indexed relicId);
+    event CreateRelic(uint256 indexed pid, address indexed to, uint256 indexed relicId);
+    event Deposit(uint256 indexed pid, uint256 amount, address indexed to, uint256 indexed relicId);
+    event Withdraw(uint256 indexed pid, uint256 amount, address indexed to, uint256 indexed relicId);
+    event EmergencyWithdraw(uint256 indexed pid, uint256 amount, address indexed to, uint256 indexed relicId);
+    event Harvest(uint256 indexed pid, uint256 amount, address indexed to, uint256 indexed relicId);
     event LogPoolAddition(
-        uint indexed pid,
-        uint allocPoint,
+        uint256 indexed pid,
+        uint256 allocPoint,
         address indexed poolToken,
         address indexed rewarder,
         address nftDescriptor,
         bool allowPartialWithdrawals
     );
-    event LogPoolModified(uint indexed pid, uint allocPoint, address indexed rewarder, address nftDescriptor);
-    event LogUpdatePool(uint indexed pid, uint lastRewardTime, uint lpSupply, uint accRewardPerShare);
-    event LogSetEmissionRate(uint indexed emissionRate);
-    event LevelChanged(uint indexed relicId, uint newLevel);
-    event Split(uint indexed fromId, uint indexed toId, uint amount);
-    event Shift(uint indexed fromId, uint indexed toId, uint amount);
-    event Merge(uint indexed fromId, uint indexed toId, uint amount);
+    event LogPoolModified(uint256 indexed pid, uint256 allocPoint, address indexed rewarder, address nftDescriptor);
+    event LogUpdatePool(uint256 indexed pid, uint256 lastRewardTime, uint256 lpSupply, uint256 accRewardPerShare);
+    event LogSetEmissionRate(uint256 indexed emissionRate);
+    event LevelChanged(uint256 indexed relicId, uint256 newLevel);
+    event Split(uint256 indexed fromId, uint256 indexed toId, uint256 amount);
+    event Shift(uint256 indexed fromId, uint256 indexed toId, uint256 amount);
+    event Merge(uint256 indexed fromId, uint256 indexed toId, uint256 amount);
 }
