+++ contracts/rewarders/MultiplierRewarderOwnable.sol
@@ -7,7 +7,7 @@ import "openzeppelin-contracts/contracts/access/Ownable.sol";
 
 /// @title Ownable extension of MultiplierRewarder that allows the owner to change the rewardMultiplier
 contract MultiplierRewarderOwnable is MultiplierRewarder, Ownable {
-    event LogRewardMultiplier(uint rewardMultiplier);
+    event LogRewardMultiplier(uint256 rewardMultiplier);
 
     /**
      * @dev Contructor called on deployment of this contract.
@@ -15,12 +15,12 @@ contract MultiplierRewarderOwnable is MultiplierRewarder, Ownable {
      * @param _rewardToken Address of token rewards are distributed in.
      * @param _reliquary Address of Reliquary this rewarder will read state from.
      */
-    constructor(uint _rewardMultiplier, address _rewardToken, address _reliquary)
+    constructor(uint256 _rewardMultiplier, address _rewardToken, address _reliquary)
         MultiplierRewarder(_rewardMultiplier, _rewardToken, _reliquary)
     {}
 
     /// @notice Set a new rewardMultiplier. Only callable by `owner`.
-    function setRewardMultiplier(uint _rewardMultiplier) external onlyOwner {
+    function setRewardMultiplier(uint256 _rewardMultiplier) external onlyOwner {
         rewardMultiplier = _rewardMultiplier;
         emit LogRewardMultiplier(_rewardMultiplier);
     }
